Organized store and test helpers
