+++ priv/react_app/src/Store/reducers/Fetching/FetchingReducer.test.js
@@ -2,9 +2,8 @@
 *   This file tests the reducers handling incoming actions.
 *   See http://redux.js.org/docs/recipes/WritingTests.html for writing action and reducer tests.
 */
-import {Actions} from './../../../Store'
+import {Actions, actionTypes} from './../../../Store'
 import reducer from './FetchingReducer'
-import actionTypes from '../../ActionTypes'
 
 describe('Fetching reducer', () => {
