--- priv/react_app/src/Test/Mocks/index.js
@@ -1,7 +1,0 @@
-import {mockLocalStorage} from './LocalStorageMock'
-import {mockHttpResponse} from './HttpResponseMock'
-
-export {
-  mockLocalStorage,
-  mockHttpResponse
-}
