+++ priv/react_app/src/TestHelper/MockLocalStorage.js
