+++ priv/react_app/src/TestHelper/index.js
@@ -1,0 +1,9 @@
+import {mockLocalStorage} from './MockLocalStorage'
+import {mockHttpResponse} from './MockHttpResponse'
+import {createMockStore} from './MockStore'
+
+export {
+  mockLocalStorage,
+  mockHttpResponse,
+  createMockStore
+}
