+++ priv/react_app/src/store/actions/Users/Users.js
@@ -1,4 +1,4 @@
-import actionTypes from './../../ActionTypes'
+import {actionTypes} from './../../../Store'
 import API from './../../../API'
 
 export const getCurrentUserStart = () => {
